Fix LNB frequency adjustment

+++ lib/libdvbapi/sec.c
@@ -137,7 +137,8 @@ int dvbfe_sec_set(struct dvbfe_handle *fe,
 		// do frequency adjustment
 		if (lof) {
 			memcpy(&localparams, params, sizeof(struct dvbfe_parameters));
-			int64_t tmpfreq = localparams.frequency - lof;
+			int tmpfreq = localparams.frequency - lof;
+
 			if (tmpfreq < 0)
 				tmpfreq *= -1;
 			localparams.frequency = (uint32_t) tmpfreq;
